requirements.txt
streamlit
pandas
pyodbc
matplotlib
python-dotenv
